+++ Tasks/CodeQuality/CodeCoverage.ps1
@@ -1,6 +1,6 @@
 <#
     .SYNOPSIS
-    Validates code coverage is over 75% for each file tested
+    Validates code coverage is over the configured minimum for each file tested
 #>
 Add-BuildTask -Name CodeCoverage -Jobs ?PesterUnit, {
     if (-not $script:pesterResult) {
@@ -11,7 +11,7 @@ Add-BuildTask -Name CodeCoverage -Jobs ?PesterUnit, {
     $script:pesterResult | ForEach-Object {
         $file = $_.Containers[0].Name
         $percent = $_.CodeCoverage.CoveragePercent
-        if ($percent -lt 75) {
+        if ($percent -lt $script:PlumberConfig.CoverageMinimum) {
             Write-Error "$percent% coverage for $file"
         }
     }