+++ Tests/Assets/Demo/DemoPipelineErr.ps1
@@ -1,1 +1,0 @@
-Add-BuildTask -Name DemoPipelineErr -Jobs Task1Err, Task2, Task3