+++ Tests/Assets/Demo/Task1.ps1
@@ -1,3 +1,0 @@
-Add-BuildTask -Name Task1 -Jobs {
-    Write-Build Yellow 'This is task 1!'
-}