+++ Tests/Assets/FailingModule/FailingModule.build.ps1
@@ -1,2 +1,0 @@
-$module = Import-Module Plumber -PassThru
-Get-ChildItem "$($module.ModuleBase)\Tasks" -Recurse -File | ForEach-Object {. $_.FullName}