+++ Tests/Assets/FailingModule/FailingModule.psm1
@@ -1,1 +1,0 @@
-﻿Get-ChildItem "$PSScriptRoot\Public", "$PSScriptRoot\Private" | ForEach-Object {. $_.FullName}