+++ Tests/Assets/PassingModule/PassingModule.build.ps1
@@ -1,6 +1,0 @@
-$module = Import-Module Plumber -PassThru
-Get-ChildItem "$($module.ModuleBase)\Tasks" -Recurse -File | ForEach-Object {. $_.FullName}
-
-Add-BuildTask -Name CodeCoverage -Jobs {
-    # Do nothing - CC not working
-}